fix extra calls from client to extension
