+++ src/client/src/containers/PageSelectFrameworks/index.tsx
@@ -3,26 +3,53 @@ import { connect } from "react-redux";
 
 import {
   EXTENSION_MODULES,
-  EXTENSION_COMMANDS
+  EXTENSION_COMMANDS,
+  FRAMEWORK_TYPE
 } from "../../utils/constants";
 
 
-import { ISelectFrameworksProps } from "./interfaces";
-import {mapStateToProps} from "./store";
+import { IStateProps, IDispatchProps } from "./interfaces";
+import {mapStateToProps, mapDispatchToProps} from "./store";
 import FrameworkCard from "./FrameworkCard";
 import styles from "./styles.module.css";
+import { getFrameworks } from "../../utils/extensionService/extensionService";
+import { parseFrameworksPayload } from "../../utils/parseFrameworksPayload";
 
-type Props = ISelectFrameworksProps;
+type Props = IStateProps & IDispatchProps;
 
-const SelectFrameworks = (props:Props) => {
-  const { frontendOptions, backendOptions } = props;
-  React.useEffect(()=>{
-    getDependencyInfoAndSetToStore();
-  },[]);
+class SelectFrameworks extends React.Component<Props> {
 
+  componentDidMount(){
+    const { frontendOptions } = this.props;
+    this.getDependencyInfoAndSetToStore();
+    if (frontendOptions.length ===0) this.getFrameworksListAndSetToStore();
+  }
 
-  const getDependencyInfoAndSetToStore = () =>{
-    const { vscode } = props;
+  getFrameworksListAndSetToStore (){
+    console.log("getFrameworksListAndSetToStore");
+    const { vscode, isPreview, setFrontendFrameworks, setBackendFrameworks } = this.props;
+
+    getFrameworks(vscode, isPreview).then((event:any)=>{
+      let message = event.data;
+      setFrontendFrameworks(
+        parseFrameworksPayload(
+          message.payload.frameworks,
+          FRAMEWORK_TYPE.FRONTEND,
+          message.payload.isPreview
+        )
+      );
+      setBackendFrameworks(
+        parseFrameworksPayload(
+          message.payload.frameworks,
+          FRAMEWORK_TYPE.BACKEND,
+          message.payload.isPreview
+        )
+      );
+    });
+  }
+
+  getDependencyInfoAndSetToStore(){
+    const { vscode } = this.props;
     // send messages to extension to check dependency info when this component loads
     vscode.postMessage({
       module: EXTENSION_MODULES.DEPENDENCYCHECKER,
@@ -40,26 +67,29 @@ const SelectFrameworks = (props:Props) => {
     });
   }
 
-  return (
-    <div>
-      <h1 className={styles.title}>Select a front-end framework</h1>
-      <div className={styles.flexContainer}>
-        {frontendOptions.map((framework) => {
-          return (
-            <FrameworkCard framework={framework} isFrontEnd={true}/>
-          );
-        })}
+  public render() {
+    const { frontendOptions, backendOptions } = this.props;
+    return (
+      <div>
+        <h1 className={styles.title}>Select a front-end framework</h1>
+        <div className={styles.flexContainer}>
+          {frontendOptions.map((framework) => {
+            return (
+              <FrameworkCard framework={framework} isFrontEnd={true}/>
+            );
+          })}
+        </div>
+        <h1 className={styles.title}>Select a back-end framework</h1>
+        <div className={styles.flexContainer}>
+          {backendOptions.map((framework) => {
+            return (
+              <FrameworkCard framework={framework} isFrontEnd={false}/>
+            );
+          })}
+        </div>
       </div>
-      <h1 className={styles.title}>Select a back-end framework</h1>
-      <div className={styles.flexContainer}>
-        {backendOptions.map((framework) => {
-          return (
-            <FrameworkCard framework={framework} isFrontEnd={false}/>
-          );
-        })}
-      </div>
-    </div>
-  );
+    );
+  }
 }
 
-export default connect(mapStateToProps)(SelectFrameworks);
+export default connect(mapStateToProps, mapDispatchToProps)(SelectFrameworks);
