+++ src/client/src/translations/whitelist_en.json
@@ -25,6 +25,7 @@
   "footer.generate",
   "footer.navAriaLabel",
   "pageAddPages.pagesTitleQuestion",
+  "pageAddPages.title",
   "pageAddPages.iconAltMessage",
   "pageCard.details",
   "azureLoginModal.azureReadMore",
